Basic Investment Calculator has been implemented. Calculation of Current value and Profit/loss are covered.
